After some cleanup and removing unused debugging code this is a first change to fix the f011/sd card write issue.
The fix changes 2 things:

1.) Fix the protocol sending sector data to the sd controller. Basically the protocol drops first 2 bytes send to the controller as part of the data_ready protocol used when executing the actual write command.
2.) Copy f011 buffer to sd controller buffer before writing (reenabled)

Still open is an issue when swap bit is used (as part of one of the later ROMs). Then the save procedure leaves sd controller in IDLE state but the f011 control in WritingSector waiting for data_ready to signal the last byte to send over. Don't know yet.