This commit as of Issue #4 fixes the bitplanes pipeline for 320/640 x 200/400 resolutions.
It fixes sprites scaling and position for 640 horizontal resolution and 400 vertical resolutions as well (not final tested).

This commit also addresses some issues with BPM mode for 640 horizontal resolution, that in the end makes it xemu compatible.

This commit keeps open that on the border area scan lines are not proportional at the moment, so working with bitplane y offset causes wrongly formated output (as in zed yago demo).