Fix issue that was tripping up Bouldermark (and likely other similar kinds of programs).
Basically shadow writes were being delayed longer than needed *and* the shadow ram primitive was being inferred as read first, which was adding even one more cycle to when writes would show up on the read port.   This tripped up Bouldermark because it was doing self-modifying code of the very next instruction and the correct data wasn't showing up when the next instruction was being fetched.

 This change changes the way shadowram/shadowrom writes are done so they happen in the current state cycle rather than being delayed, and also fixes the shadow ram primitive to be write first mode so that if you write a value to address A on cycle 0 and read from it on the same cycle (or next cycle) you get the data just written.