One tweak to the monitor state capture stuff so it captures everything as it is on the current cycle (including the address being driven on the current cycle), and some minor changes to the FMC debug output on the Nexus video board so I can keep tabs on what's going on in real time if needed.
There's something on the hairy edge of timing that the debug output is effecting on I haven't been able to nail down yet.   This particular combination works, but the prior stuff (which is older) no longer does.   In theory this should have any effect on the other boards though since these signals aren't connected anywhere on the outside and thus ought to get optimized away anyway.