Switch back to using a tristate value instead of a '1' and a wire-and for IRQs, but use 'H' instead of 'Z' so the unproven value becomes 1 and not 0.
Also, use 'H' for external keyboard related tristates so that if no KB hardware is connected that we get 1s instead of 0s for the physical keyboard matrix.   I don't think this should have any effect on the real external keyboard hardware but Paul should try to confirm on his boards.