Experimental rework of sprite block.   Probably needs more testing but I think boulder mark is working.
The basic idea was to not expand the data to 128 bits (and have multiple copies) but instead just do shifts of different widths from the original 64-bit storage, and use a little bit of combinatorial logic to differentiate between MC mode and Mono mode.    Reduces each sprite instance by a few hundred cells.