Experimental rewrite of VHDL UART monitor.
This commit swaps out the existing VHDL UART monitor implementation for an entirely new implementation using a combination of a 65[C]02 core of my own design, some custom support hardware for supporting the existing monitor hardware interface, and the software to drive the whole thing.  The main motivation was saving resources in the FPGA.  Some notable highlights:

1) The 6502/65C02 core is one I wrote earlier this year as mostly as an exercise to see if it would be possible to put the microcode into block RAMs instead of normal LUTs to save logic resources.   Luckily it worked out and the design winds up being pretty compact, as well as also very easy to modify/understand.  It provides signals for interfacing to both synchronous and asynchronous memories.   It should be cycle accurate with both the 6502 and 65C02 depending on which way it is built.  It even shares some bugs with the original 6502.   One thing NOT implemented yet are any "illegal" opcodes from the NMOS 6502.   The core passes Klaus Dormann's entire 6502/65C02 test suite (well, everything but the WDC STP and WAI instructions, which I didn't implement).

2) The monitor controller implements all of the things that are best done in hardware including the serial I/O, monitor control functions, state and history recording, etc.   The hardware uses dual ported block rams to store the CPU history and recent states.

3) The "computer" for the monitor consists of the CPU, a 2K RAM, a 2K ROM, the monitor controller, the serial ports, and the block RAMs.   The 2K ROM is nearly full but should be pretty feature complete relative to the original.   The 2K RAM has about 1.5K unused at the moment that could be used to store more monitor code without much effort.

The new monitor uses about 25% of the logic cell count of the original. It does use a few more block RAMs than the original, mostly for the 6502 microcode and the monitor RAM/ROM.   It's possible for the 6502 core to not use a block RAM for the microcode at the expense of more logic resources by editing the 6502_ucode.v file to not force the use of block rams.  It actually makes it run faster, but my original goal was for minimizing logic and routing resources.  YMMV.

Here's the rub: The CPU core and monitor core are written in Verilog.   I'm just not a VHDL guy, and it'd be much harder to have done the CPU core the way I did in VHDL since I made extensive use of the preprocessor for the microcode definitions.   Fortunately it's rather easy to mix languages in an FPGA design.   At the end of the day, it's all just wires.