Combine 2K RAM and 2K ROM and add 65CE02 disassembler.
Rather than have separate memories, just combine them and map the same block RAM to both high and low memory.  It requires fewer routing resources anyway (less muxing of the CPU data lines) and I can still have the monitor bus controller write protect everything but zero page and the stack very easily.

There's still about 1/2K of unused space left for other monitor features.   Also, the disassembler tables (in particular the opcode metadata table) has a lot of redundancy in it.  It could probably be rewritten to be derived via code and use <256 bytes of space.

One possible idea for the remaining code space would be to take over the terminal emulation from the matrix compositor and just do it in software in the monitor.   Then the matrix compositor could be reduced to basically just the character fetch and scanout logic.