Monitor additions/fixes.
Fix up some bugs with register state output.  The MAPH state was wrong (wrong byte used).   I also found that having L then H was driving me crazy so I swapped them. ;)
Added some additional state to show various CPU/VIC ROM mapping controls in the state output.
Fixed disassembly so it uses full 32-bit addresses and it's forced to the CPU view.  Now you can disassemble anything with a system memory address like the C65 monitor can.
Added the ability for the disassembly to be halted as soon as a key is detected on the serial port so if you so something silly you can break out of it.  Should probably add this to all of the other looping commands too.