implement improvements to USE_LOCAL_CC65, #274
this reverses the logic around the use of the ifdef/ifeq of USE_LOCAL_CC65,
as discussed in #220

also adds the 'DEPEND'-on variable, and apply this to all targets that use the cc65-tools

assumes that ca65,ld65,cl65 are available when cc65 is available.
ie: a target using 'ld65' will depend on if CC65_DEPEND is set or not.