fixing version.sh script, REF should now be in working tree, #291
error was:
fatal: ambiguous argument 'ghdl-crash-20200604..138-hdmi-audio-27mhz': unknown revision or path not in the working tree.
so now we are looking for a commit directly on the branch (4957c7ba).
NOTE that when this branch is moved to master (or similar), it may be that
the commit above (that we are looking for (to make the incrementing number))
will not be found, causing FATAL again.

also removed a debug printline