Return the file descriptor in A from trap_dos_openfile (#491)
* Return the file descriptor in A from trap_dos_openfile.

trap_dos_openfile opens a file descriptor for the file and it's the programmer's responsibility to close it. In order to close it, they need to know what it is.

The A register was chosen to align with trap_dos_opendir and trap_dos_findfirst.

* Introduce return_from_trap_with_success_and_file_descriptor_in_a.

Have trap_dos_opendir, trap_dos_openfile and trap_dos_findfirst use it, in order to save some bytes.