fix: make build version consistent
The version name of the file was generated after the build finished,
where as the version inside the files is created at the start of the
build. This could result in the date being different if the hour/day
changed while building.

This change stores the filename version inside src/version.txt, which
is then used to rename the files after the builds are done, fixing this
problem.