fix(viciv): setting HOTREG to 0 will clear recalc flag #714
This allows to clear changes to HOTREGs.

Normally when you set HOTREG to 0 any change to a hot register after
this will set the reclac flag. Reenabling HOTREG by setting it to 1
will then trigger a recalculation.

By setting HOTREG first to 0 and then to 1, you are able to clear the
recalc flag, thus avoiding the recalculation.

This is needed by ROM.