Repair mouse and paddle in port 2
The CIA routing of peripheral port analog pins to SID1 was incorrectly wired to CIA PORTB, when it should have been PORTA. Also, the SID selection was backwards: SID1 is "potr", peripheral port 1 is "pota".

I double-checked that we want to label SID1+2 as the "right" SIDs, so swapping "potl" for "potr" is the correct change.