fix(megaflash): possible fix for bootloop problem - wait! #683
The bootloop seems to be a timing issue: when more slots are filled,
the time to scan them actaully gets smaller, as detecting an EMPTY slot
takes the most time. Adding 1-2 frames of waiting before the reconfigure
seems to fix the problem.

Also moved wait for keyrelease into mhexes as mhx_until_keys_released