Improve SDRAM constraints
This fix forces the I/O registers of the SDRAM controller to be placed
inside the I/O block (rather than in the main fabric of the FPGA).

The key is the "attribute iob".

However, in order to get full benefit, it's necessary to have explicit
registers of the "output enable", or rather "tristate enable".

Furthermore, the constraints are no longer necessary, because there is
nothing more to constrain.